[breadboard-cli] start of refactor
